Renamed ConditionalBlockWrapper file to ConditionalBlock
